Fixed the main test to use the snapcatch2 main.
Now it works properly with `./mk -t`.

Cleaned up the order in which the headers are included in the tests to allow
for some verification of some headers (i.e. that they can be included first).

Moved the Doxygen check in the doc/CMakeLists.txt file.

Fixed the tests that verified the "#id" not appearing at the start because
that was wrong anyway (there are cases where it can appear somewhere else
and it is valid). I fixed that a while back, but not the tests.