Got the coverage test to work.
Since I converted to the ./mk -c option, I had not had this project working
properly just yet. Now it generates the coverage as expected and in the
same place as the other projects (lcov.snapwebsites.org).

Fixed memory leaks in a compiler test.

Added a few more lines to ignore (because they are never reached).